Preview custom icon before saving.
Also, use the custom icon in page headers, where appropriate.

Fixes #1154
