+++ static/index.html
@@ -19,6 +19,8 @@
         <div id="thing-title-back-flex"></div>
         <div id="thing-title-container">
           <img id="thing-title-icon" alt="Thing Title" src="#" />
+          <webthing-custom-icon id="thing-title-custom-icon" class="hidden">
+          </webthing-custom-icon>
           <span id="thing-title-name"></span>
         </div>
         <div id="thing-title-speech-flex"></div>
@@ -355,21 +357,26 @@
       <button id="context-menu-back-button" class="back-button icon-button"></button>
       <div id="context-menu-heading">
         <img id="context-menu-heading-icon" alt="Thing Title" src="#" />
+        <webthing-custom-icon id="context-menu-heading-custom-icon" class="hidden">
+        </webthing-custom-icon>
         <span id="context-menu-heading-text"></span>
       </div>
       <div id="context-menu-content">
         <div id="context-menu-content-edit" class="hidden">
-          <div id="edit-thing-icon"></div>
+          <div id="edit-thing-icon">
+            <webthing-custom-icon id="edit-thing-custom-icon" class="hidden">
+            </webthing-custom-icon>
+          </div>
           <div id="edit-thing-metadata">
             <input type="text" id="edit-thing-name">
             <select id="edit-thing-type"></select>
             <span id="edit-thing-spacer"></span>
-            <label for="edit-thing-custom-icon"
+            <label for="edit-thing-custom-icon-input"
               id="edit-thing-custom-icon-label" class="edit-thing-button text-button hidden">
               Choose icon...
             </label>
-            <input type="file" class="edit-thing-custom-icon hidden"
-              id="edit-thing-custom-icon"
+            <input type="file" class="edit-thing-custom-icon-input hidden"
+              id="edit-thing-custom-icon-input"
               accept="image/jpeg,image/png,image/svg+xml">
             <span id="edit-thing-label" class="hidden"></span>
           </div>
