+++ static/js/app.js
@@ -213,33 +213,33 @@ RuleScreen = require('./rule-screen');
 Speech = require('./speech');
 
 // load web components
-require('./components/action');
-require('./components/base-component');
-require('./components/binary-sensor-capability');
-require('./components/boolean-property');
-require('./components/brightness-property');
-require('./components/color-control-capability');
-require('./components/color-property');
-require('./components/color-temperature-property');
-require('./components/current-property');
-require('./components/custom-capability');
-require('./components/energy-monitor-capability');
-require('./components/frequency-property');
-require('./components/instantaneous-power-property');
-require('./components/label-capability');
-require('./components/label-property');
-require('./components/level-property');
-require('./components/light-capability');
-require('./components/multi-level-sensor-capability');
-require('./components/multi-level-switch-capability');
-require('./components/number-property');
-require('./components/on-off-property');
-require('./components/on-off-switch-capability');
-require('./components/slider-property');
-require('./components/smart-plug-capability');
-require('./components/string-property');
-require('./components/switch-property');
-require('./components/voltage-property');
+require('./components/capability/binary-sensor');
+require('./components/capability/color-control');
+require('./components/capability/custom');
+require('./components/capability/energy-monitor');
+require('./components/capability/label');
+require('./components/capability/light');
+require('./components/capability/multi-level-sensor');
+require('./components/capability/multi-level-switch');
+require('./components/capability/on-off-switch');
+require('./components/capability/smart-plug');
+require('./components/icon/custom');
+require('./components/property/action');
+require('./components/property/boolean');
+require('./components/property/brightness');
+require('./components/property/color');
+require('./components/property/color-temperature');
+require('./components/property/current');
+require('./components/property/frequency');
+require('./components/property/instantaneous-power');
+require('./components/property/label');
+require('./components/property/level');
+require('./components/property/number');
+require('./components/property/on-off');
+require('./components/property/slider');
+require('./components/property/string');
+require('./components/property/switch');
+require('./components/property/voltage');
 
 if (navigator.serviceWorker) {
   navigator.serviceWorker.register('/service-worker.js', {
