+++ static/js/components/property/on-off.js
@@ -9,7 +9,7 @@
  */
 'use strict';
 
-const SwitchProperty = require('./switch-property');
+const SwitchProperty = require('./switch');
 
 class OnOffProperty extends SwitchProperty {
   connectedCallback() {
