+++ static/js/components/property/slider.js
@@ -1,7 +1,7 @@
 /**
- * ColorProperty
+ * SliderProperty
  *
- * A bubble showing a color input.
+ * A bubble showing a slider.
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,7 +9,7 @@
  */
 'use strict';
 
-const BaseComponent = require('./base-component');
+const BaseComponent = require('../base-component');
 
 const template = document.createElement('template');
 template.innerHTML = `
@@ -22,7 +22,7 @@ template.innerHTML = `
       font-size: 1.6rem;
     }
 
-    .webthing-color-property-container {
+    .webthing-slider-property-container {
       width: 10rem;
       height: 10rem;
       border-radius: 5rem;
@@ -31,14 +31,25 @@ template.innerHTML = `
       position: relative;
     }
 
-    .webthing-color-property-contents {
+    .webthing-slider-property-contents {
       position: absolute;
       top: 50%;
       left: 50%;
       transform: translate(-50%, -50%);
     }
 
-    .webthing-color-property-name {
+    .webthing-slider-property-form {
+      width: 8rem;
+      height: 8rem;
+    }
+
+    .webthing-slider-property-slider {
+      width: 8rem;
+      margin: 0;
+      transform: rotate(-90deg) translate(-3rem, 0);
+    }
+
+    .webthing-slider-property-name {
       text-align: center;
       max-width: 10rem;
       overflow-wrap: break-word;
@@ -46,27 +57,34 @@ template.innerHTML = `
       display: inline-block;
     }
   </style>
-  <div id="container" class="webthing-color-property-container">
-    <div id="contents" class="webthing-color-property-contents">
-      <input type="color" id="color" class="webthing-color-property-color">
+  <div id="container" class="webthing-slider-property-container">
+    <div id="contents" class="webthing-slider-property-contents">
+      <form id="form" class="webthing-slider-property-form">
+        <input type="range" id="slider" class="webthing-slider-property-slider">
+      </form>
     </div>
   </div>
-  <div id="name" class="webthing-color-property-name"></div>
+  <div id="name" class="webthing-slider-property-name"></div>
 `;
 
-class ColorProperty extends BaseComponent {
+class SliderProperty extends BaseComponent {
   constructor() {
     super(template);
 
-    this._input = this.shadowRoot.querySelector('#color');
+    this._input = this.shadowRoot.querySelector('#slider');
     this._name = this.shadowRoot.querySelector('#name');
 
     this._onChange = this.__onChange.bind(this);
   }
 
   connectedCallback() {
-    this.name = this.dataset.name;
+    if (!this.name) {
+      this.name = this.dataset.name;
+    }
 
+    this._upgradeProperty('min');
+    this._upgradeProperty('max');
+    this._upgradeProperty('step');
     this._upgradeProperty('value');
     this._upgradeProperty('disabled');
 
@@ -75,6 +93,30 @@ class ColorProperty extends BaseComponent {
 
   disconnectedCallback() {
     this._input.removeEventListener('change', this._onChange);
+  }
+
+  get min() {
+    return this._input.min;
+  }
+
+  set min(value) {
+    this._input.min = value;
+  }
+
+  get max() {
+    return this._input.max;
+  }
+
+  set max(value) {
+    this._input.max = value;
+  }
+
+  get step() {
+    return this._input.step;
+  }
+
+  set step(value) {
+    this._input.step = value;
   }
 
   get value() {
@@ -121,5 +163,5 @@ class ColorProperty extends BaseComponent {
   }
 }
 
-window.customElements.define('webthing-color-property', ColorProperty);
-module.exports = ColorProperty;
+window.customElements.define('webthing-slider-property', SliderProperty);
+module.exports = SliderProperty;
