+++ static/js/new-web-thing.js
@@ -36,6 +36,8 @@ class NewWebThing {
     this.thingType = this.element.getElementsByClassName('new-thing-type')[0];
     this.customIcon =
       this.element.getElementsByClassName('new-thing-custom-icon')[0];
+    this.customIconInput =
+      this.element.getElementsByClassName('new-thing-custom-icon-input')[0];
     this.customIconLabel =
       this.element.getElementsByClassName('new-thing-custom-icon-label')[0];
     this.label =
@@ -53,8 +55,8 @@ class NewWebThing {
       this.element.getElementsByClassName('new-web-thing-save-button')[0];
     this.element.addEventListener('click', this.handleClick.bind(this));
     this.thingType.addEventListener('change', this.handleTypeChange.bind(this));
-    this.customIcon.addEventListener('change',
-                                     this.handleIconUpload.bind(this));
+    this.customIconInput.addEventListener('change',
+                                          this.handleIconUpload.bind(this));
   }
 
   /**
@@ -65,7 +67,10 @@ class NewWebThing {
     return `
       <div id="new-web-thing-${Utils.escapeHtmlForIdClass(this.originalId)}"
            class="new-thing web-thing">
-        <div class="new-thing-icon"></div>
+        <div class="new-thing-icon">
+          <webthing-custom-icon class="new-thing-custom-icon hidden">
+          </webthing-custom-icon>
+        </div>
         <div class="new-thing-metadata">
           <input type="text" class="new-web-thing-url"
             placeholder="Enter web thing URL"></input>
@@ -78,8 +83,8 @@ class NewWebThing {
             Choose icon...
           </label>
           <span class="new-web-thing-label">Web Thing</span>
-          <input type="file" class="new-thing-custom-icon hidden" id="${id}"
-            accept="image/jpeg,image/png,image/svg+xml">
+          <input type="file" class="new-thing-custom-icon-input hidden"
+            id="${id}" accept="image/jpeg,image/png,image/svg+xml">
           <div class="new-web-thing-controls">
             <button class="new-web-thing-cancel-button text-button">
               Cancel
@@ -107,6 +112,7 @@ class NewWebThing {
       this.thingType.options[this.thingType.selectedIndex].value;
 
     this.customIconLabel.classList.add('hidden');
+    this.customIcon.classList.add('hidden');
 
     let cls = '';
     switch (capability) {
@@ -136,6 +142,7 @@ class NewWebThing {
         break;
       case 'Custom':
         this.customIconLabel.classList.remove('hidden');
+        this.customIcon.classList.remove('hidden');
         break;
       default:
         break;
@@ -158,11 +165,11 @@ class NewWebThing {
   handleIconUpload() {
     this.label.classList.add('hidden');
 
-    if (this.customIcon.files.length === 0) {
+    if (this.customIconInput.files.length === 0) {
       return;
     }
 
-    const file = this.customIcon.files[0];
+    const file = this.customIconInput.files[0];
     if (!['image/jpeg', 'image/png', 'image/svg+xml'].includes(file.type)) {
       this.label.innerText = 'Invalid file.';
       this.label.classList.add('error');
@@ -186,6 +193,9 @@ class NewWebThing {
         data: btoa(e.target.result),
       };
       this.saveButton.disabled = false;
+
+      this.customIcon.iconHref = URL.createObjectURL(file);
+      this.customIcon.classList.remove('hidden');
     };
 
     this.saveButton.disabled = true;
